RunSafely/lit: Allow separate stdout/stderr redirection.
This will be necessary for the upcoming SPEC support for cmake/lit which will
use the reference data coming with SPEC instead of the custom one in
the Externals directory.

This commit also reverses the setting for the RunSafely output format to
produce the traditional output format by default because that is still
used by most benchmarks. That way fewer lit.local.cfg files are
necessary.